Moved files out of `shortNameResolver`
