+++ formver.compiler-plugin/core/src/org/jetbrains/kotlin/formver/core/names/NameType.kt
@@ -1,6 +1,5 @@
 package org.jetbrains.kotlin.formver.core.names
 
-import org.jetbrains.kotlin.formver.core.names.shortNameResolver.nameOnlyCandidates
 import org.jetbrains.kotlin.formver.viper.AnyName
 import org.jetbrains.kotlin.formver.viper.CandidateName
 import org.jetbrains.kotlin.formver.viper.NameTypeBase
