+++ formver.compiler-plugin/core/src/org/jetbrains/kotlin/formver/core/names/ScopedName.kt
@@ -6,8 +6,6 @@
 package org.jetbrains.kotlin.formver.core.names
 
 
-import org.jetbrains.kotlin.formver.core.names.shortNameResolver.buildCandidates
-import org.jetbrains.kotlin.formver.core.names.shortNameResolver.nameWithDependentPrefixCandidates
 import org.jetbrains.kotlin.formver.viper.AnyName
 import org.jetbrains.kotlin.formver.viper.CandidateName
 import org.jetbrains.kotlin.formver.viper.NameTypeBase
