+++ formver.compiler-plugin/core/src/org/jetbrains/kotlin/formver/core/names/ViperKeyword.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.kotlin.formver.core.names.shortNameResolver
+package org.jetbrains.kotlin.formver.core.names
 
 import org.jetbrains.kotlin.formver.viper.AnyName
 import org.jetbrains.kotlin.formver.viper.CandidateName
